requirements.txt
streamlit
PyMuPDF
sentence-transformers
google-generativeai
python-dotenv
chromadb